Added the customMatchingFunction option for the ProgramRules.
Made the LoggerPlugin output more information on program rule matching.

Use EnumDesktopWindows instead of EnumWindows. Should be pretty much the same.

Use HashMultiSet again for the hidden windows. This seems to be the only way to make it work correctly when workspaces are changed really rapidly.

Hiding of windows is done now synchronously, as is showing, with DeferWindowPos.

Fixed a race condition in GetWindowSmallIconAsBitmap.